# specs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Checkmate API
  version: 1.0.0
  summary: Self-hosted personal task manager.
  description: |
    A single-user task manager organised around four parallel **contexts**
    (Upsun, Personal, Gaal, Arkea): capture without choosing a context, triage
    deliberately, and read one daily brief.

    ## Authentication

    Three credential kinds, all resolving to the same user:

    | Kind | Used by | Obtained |
    |---|---|---|
    | Session cookie | Web | Google OIDC via `/auth/login/google` |
    | Device token (`cm_…`) | iOS, macOS, CLI | `POST /v1/tokens` (session only) or the CLI |
    | OAuth access token (`cmat_…`) | Remote MCP clients | `/oauth/authorize` + `/oauth/token` |

    Scopes are `read` (all GETs) and `write` (all mutations). A session cookie
    always carries both.

    ## Conventions

    - Collections return `{"data": [...], "next_cursor": "<id>"|null}`; single
      resources return the bare object.
    - Ids are UUIDv7 strings. Because v7 sorts chronologically, `id DESC` is
      creation order, which is how every list is ordered.
    - Timestamps are RFC3339 UTC (`2026-07-25T14:03:11.482Z`). Calendar dates are
      plain `YYYY-MM-DD` with no timezone.
    - Estimates are whole minutes.
    - Errors are `{"error": "...", "fields": {"due_on": "..."}}`. When `fields` is
      present, render each entry against its control rather than as a toast.

    ## Ordering

...
    opens with undated tasks.

...
  contact:
    name: Checkmate
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /v1/tasks:
    get:
      tags: [Tasks]
      summary: List and filter tasks
      description: |
        `context_id=null`, `project_id=null` and `parent_id=null` select rows where
        that column is NULL. `parent_id=null` and `top_level=true` are the same
        question.
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/includeDeleted"
        - name: sort
          in: query
          description: |
            Column to order by. Rows with no value for it sort last in both
            directions.
          schema:
            type: string
            default: created_at
            enum: [created_at, updated_at, due_on, planned_on, title, estimate_minutes, completed_at, status]
        - name: order
          in: query
          description: |
            Direction. Defaults to `desc` for `created_at` — newest first, which is
            what a task list wants — and `asc` for every other field, so the
            soonest date or shortest estimate leads.
          schema:
            type: string
            enum: [asc, desc]
        - name: status
          in: query
          description: Repeatable, or comma-separated.
          schema:
            type: array
            items: { $ref: "#/components/schemas/TaskStatus" }
          style: form
          explode: true
        - name: kind
          in: query
          description: Repeatable, or comma-separated. Derived server-side.
          schema:
            type: array
            items: { $ref: "#/components/schemas/TaskKind" }
          style: form
          explode: true
        - name: context_id
          in: query
          description: A context id, or the literal `null` for the inbox.
          schema: { type: string }
        - name: project_id
          in: query
          description: A project id, or the literal `null`.
          schema: { type: string }
        - name: parent_id
          in: query
          description: A task id, or the literal `null` for top-level tasks only.
          schema: { type: string }
        - name: top_level
          in: query
          description: Exclude subtasks. Equivalent to `parent_id=null`.
          schema: { type: boolean }
        - name: delegated_to_id
          in: query
          schema: { type: string }
        - name: recurrence_id
          in: query
          description: Every occurrence spawned from one series.
          schema: { type: string }
        - name: blocked_by_id
          in: query
          description: |
            The reverse edge: which tasks are waiting on this one. `null` selects
            tasks with no blocker at all.
          schema: { type: string }
        - name: planned_on
          in: query
          schema: { $ref: "#/components/schemas/Date" }
        - name: planned_before
          in: query
          description: Inclusive.
          schema: { $ref: "#/components/schemas/Date" }
        - name: planned_after
          in: query
          description: Inclusive.
          schema: { $ref: "#/components/schemas/Date" }
        - name: due_on
          in: query
          schema: { $ref: "#/components/schemas/Date" }
        - name: due_before
          in: query
          description: Inclusive.
          schema: { $ref: "#/components/schemas/Date" }
        - name: due_after
          in: query
          description: Inclusive.
          schema: { $ref: "#/components/schemas/Date" }
        - name: q
          in: query
          description: |
            Substring match over title and details. Wildcards typed by the user are
            escaped and match literally.
          schema: { type: string }
      responses:
        "200":
          description: OK.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: |
            Validation failed. `sort` or `order` for an unknown value; `cursor`
            when it was issued for a different sort.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      tags: [Tasks]
      summary: Create a task
      description: |
        Only `title` is required. Omitting `context_id` captures into the inbox,
        which is the intended fast path: `status` then defaults to `inbox`, and to
        `todo` when a context is given.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreate" }
      responses:
        "201":
          description: Created. `Location` names the new task.
          headers:
            Location:
              schema: { type: string, examples: ["/v1/tasks/019f98..."] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A device token (`cm_…`) or an OAuth access token (`cmat_…`). An OAuth token
        is additionally checked to have been issued for this resource. A refresh
        token presented here is refused.
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-checkmate_session
      description: |
        Set by the sign-in callback. `HttpOnly`, `Secure`, `SameSite=Lax`. The
        `__Host-` prefix requires `Secure`, so a development server over plain
        http uses the unprefixed name `checkmate_session` instead.

        Cookie-authenticated mutations must additionally prove same origin, via
        `Sec-Fetch-Site` or a matching `Origin`; a request with neither is refused.
    oauth2:
      type: oauth2
      description: For remote MCP clients. PKCE with S256 is mandatory.
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          refreshUrl: /oauth/token
          scopes:
            read: See tasks, projects, contexts and people.
            write: Create, change and delete.
            offline_access: Receive a refresh token.
  parameters:
...
    limit:
      name: limit
      in: query
      description: Page size.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from the previous page. **Opaque — do not parse it.**

        On a sorted task listing it is a keyset position bound to that exact sort,
        so reusing it with a different `sort` or `order` is a 422 rather than a
        quietly wrong page.
      schema: { type: string }
    includeDeleted:
      name: include_deleted
      in: query
      description: |
        Include tombstoned rows, which carry `deleted_at`. There is no undelete
        endpoint, so these are historical only.
      schema: { type: boolean, default: false }
  responses:
    BadRequest:
      description: |
        Malformed JSON, an unknown field, or more than one JSON document. Unknown
        fields are rejected rather than ignored, so a misspelled key on PATCH is
        reported instead of looking like a successful no-op.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        No credential, or one that is unknown, revoked, expired, or issued for a
        different resource. `WWW-Authenticate` names the protected resource
        metadata document and the scopes the request needs.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        The credential lacks the scope this method needs — `read` for GET, `write`
        for a mutation — or a cookie-authenticated mutation failed the same-origin
        check. A scope failure carries `WWW-Authenticate` with
        `error="insufficient_scope"`, which is what lets an OAuth client step up.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: Over 1 MiB. Only reachable through `details`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UnsupportedMediaType:
      description: "`Content-Type` is not JSON."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: |
        Well-formed but not applicable. `fields` maps each offending input to a
        message; render them inline rather than as one toast.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            task:
              value:
                error: validation failed
                fields:
                  due_on: must be a YYYY-MM-DD date
                  title: is required
  schemas:
    Date:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
      description: |
        A calendar date with no timezone. Validated as a real date, so
        `2026-02-31` is rejected rather than merely shape-checked.
      examples: ["2026-08-15"]
    Timestamp:
      type: string
      format: date-time
      description: RFC3339, UTC, milliseconds.
      examples: ["2026-07-25T14:03:11.482Z"]
    Rev:
      type: integer
      format: int64
      description: |
        A global change counter, unique across every table and monotonic. Used as
        the sync cursor. Not writable.
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        fields:
          type: object
          description: Per-input messages, keyed by the request field name.
          additionalProperties: { type: string }
...
    TaskStatus:
      type: string
      enum: [inbox, todo, in_progress, blocked, delegated, done, cancelled]
      description: |
        Exhaustive; enforced by a database CHECK. `inbox` means captured but not
        triaged. `delegated` requires `delegated_to_id`, in both directions.
        `blocked` does **not** require `blocked_by_id`, because a task can be
        blocked by something outside Checkmate. `cancelled` is not `done`.
    TaskKind:
      type: string
      enum: [short, long, recurring, delegated, blocked]
      readOnly: true
      description: |
        Derived by the server and never stored, so it cannot go stale. Precedence
        is strict: `recurring` if `recurrence_id` is set, else `delegated` if
        `delegated_to_id` is set, else `blocked` if `blocked_by_id` is set or the
        status is `blocked`, else `long` if it has a live child, else `short`.
    CaptureMethod:
      type: string
      enum: [form, api, hermes, chrome_ext, ios_widget, voice, recurrence]
      description: |
        How the task entered Checkmate — an axis independent of `source`, which is
        where it came from. Set on create, then read-only. `hermes` is what the MCP
        endpoint stamps, `recurrence` what the spawner stamps. Diagnostic: show it
        in a detail footer, not in a list row.
...
    Task:
      type: object
      required:
        [id, title, status, kind, capture_method, created_at, updated_at, rev]
      properties:
        id: { type: string }
        context_id:
          type: ["string", "null"]
          description: Null means the task is in the inbox, awaiting triage.
        project_id: { type: ["string", "null"] }
        parent_id:
          type: ["string", "null"]
          description: Subtasks nest to any depth. Cycles are rejected at any depth.
        recurrence_id:
          type: ["string", "null"]
          readOnly: true
          description: Set on an occurrence spawned from a series. Not detachable.
        occurrence_on:
          oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }]
          readOnly: true
          description: The date this occurrence stands for.
        source:
          type: ["string", "null"]
          description: A `Source.key`.
        capture_method: { $ref: "#/components/schemas/CaptureMethod" }
        title: { type: string, minLength: 1, maxLength: 500 }
        details:
          type: ["string", "null"]
          description: Plain text. Counts against the 1 MiB body limit.
        status: { $ref: "#/components/schemas/TaskStatus" }
        due_on:
          oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }]
          description: The deadline that exists in the world.
        planned_on:
          oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }]
          description: |
            The day the user intends to work on it. Distinct from `due_on` — a task
            can be due Friday and planned for Tuesday. Conflating them destroys the
            model.
        estimate_minutes:
          type: ["integer", "null"]
          format: int64
          minimum: 1
          description: Whole minutes, greater than zero.
        delegated_to_id: { type: ["string", "null"] }
        blocked_by_id: { type: ["string", "null"] }
        reference_url:
          type: ["string", "null"]
          description: |
            Absolute http or https URL pointing at the origin — a Slack permalink,
            an email thread. The only external pointer; there are no attachments.
        reference_label: { type: ["string", "null"] }
        kind: { $ref: "#/components/schemas/TaskKind" }
        completed_at:
          oneOf: [{ $ref: "#/components/schemas/Timestamp" }, { type: "null" }]
          readOnly: true
          description: Server-managed. Cleared automatically when a task is reopened.
        cancelled_at:
          oneOf: [{ $ref: "#/components/schemas/Timestamp" }, { type: "null" }]
          readOnly: true
        created_at: { $ref: "#/components/schemas/Timestamp" }
        updated_at: { $ref: "#/components/schemas/Timestamp" }
        deleted_at:
          oneOf: [{ $ref: "#/components/schemas/Timestamp" }, { type: "null" }]
          readOnly: true
          description: Present only on a tombstone.
        rev: { $ref: "#/components/schemas/Rev" }
    TaskCreate:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        details: { type: ["string", "null"] }
        context_id:
          type: ["string", "null"]
          description: Omit to capture into the inbox.
        project_id:
          type: ["string", "null"]
          description: Must belong to the same context as `context_id`.
        parent_id: { type: ["string", "null"] }
        source: { type: ["string", "null"] }
        capture_method:
          $ref: "#/components/schemas/CaptureMethod"
          default: api
        status:
          $ref: "#/components/schemas/TaskStatus"
          description: |
            Defaults to `inbox` without a context, `todo` with one. Setting
            `delegated` requires a delegate.
        due_on: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
        planned_on: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
        estimate_minutes: { type: ["integer", "null"], format: int64, minimum: 1 }
        delegated_to_id: { type: ["string", "null"] }
        delegated_to:
          type: ["string", "null"]
          description: |
            A person's **name**, resolved or created in the same call. Mutually
            exclusive with `delegated_to_id`. Create-only; not accepted on PATCH.
        blocked_by_id: { type: ["string", "null"] }
        reference_url: { type: ["string", "null"] }
        reference_label: { type: ["string", "null"] }
...
    TaskList:
      type: object
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Task" }
        next_cursor: { type: ["string", "null"] }
...